Create README.md
This project involves the development of a Python-based program for Tamarit Estudios, a company in Valencia, Spain, that automates the extraction of specific data from PDF invoices. The program efficiently parses invoices from various energy companies, extracting crucial fields such as CUPS, total consumption, and supply address. It then organizes this data into tables for easy viewing and further processing through an HTML interface.

The program is integrated with the DataDis API, ensuring compliance with industry standards in the energy sector. By automating this process, the project has significantly reduced the time and effort required for manual data entry, leading to improved accuracy and considerable cost savings for Tamarit Estudios.